Move change_email/start to password-protected endpoint
