+++ packages/backend/src/services/web/UserProtectedEndpointsService.js
@@ -82,7 +82,11 @@ class UserProtectedEndpointsService extends BaseService {
         });
 
         Endpoint(
-            require('../../routers/user-protected/change-password.js')
+            require('../../routers/user-protected/change-password.js'),
+        ).attach(router);
+
+        Endpoint(
+            require('../../routers/user-protected/change-email.js'),
         ).attach(router);
     }
 }
